refactor to `_annotate`
